Add runtime check for out-of-bounds CUDA dimension
The CUDA runtime system and driver provide very little in the
way of helpful or clear error messages. Especially in the case of
kernel dimensions being out-of-bound. Depending on what CUDA-CC
one uses (e.g. 2.0, 3.5, 5.0, etc.) there is a max-limit to how
large the X, Y, and Z dimensions can be for the kernel.

This patch adds a check which will error out if we generate a dimension
greater than permitted by the CUDA-CC.

Furthermore a test is provided.