Added support for HWLOC-CUDA
This involves a major re-arranging of HWLOC stuff. In particular,
the MT_HWLOC stuff has been moved out into it's own libsac file,
at least all the stuff that is generic - MT speicifc stuff is still
in the hwloc_data.c file.

Additionally a libsac cuda file was added which does some CUDA
specific stuff.

And finally the SETUP/FINALIZE ICMs have been updated, and new ones
added to account for the changes.

As it currently stands, it is assumed that MT uses one mode of HWLOC
while CUDA uses another - testing shows that this works fine and
existing functionality has been maintained. In the case of CUDA_hybrid
or otherwise, this is not clear.

No new flags have been added, this and further checking is left
for later work.