Minor Fix to pragma cudalinkname
Sadly there is nothing to explain what this pragma is for. Furthermore,
other than then the scanparser and resolvepragma, there is no further
implementation of it.

My belief is that this pragma is meant to be used to make use of
CUDA specific API functions, by altering the linkname (similar to
the linkname pragma) - in particular this is useful for WL which
are cudarizable, as one can make use of CUDA device functions
which *may* offer some performance benefit.

This fix just resolves in issue the wrong variable being accessed,
resulting in a compiler crash.