Add new pragma `header'
The purpose of this pragma is to specify that an external function
declaration need not have a prototype, as this is provided in a
separate header. This is particular useful/important (?) in the case
of CUDA, which in the case of math functions provides its own library.
The prototypes declarations are somewhat different from the math.h ones
in LIBC, causing NVCC to error out.

A basic example of usage is:

  external double sin(double x);
     #pragma linksign [0,1]
     #pragma header "<math.h>"

The argument of `header' is the header file that contains the prototype
of the function declaration. This value is stored within the n_module
node as an attribute. Other instances of the pragma are stored here
as well as a StringSet. The function declaration is marked as being
provided externally, which disables its prototype from being generated.
The n_module attribute elements are printed out during code generation.

Of course custom header files can be used as well, e.g.
"\"myheader.h\"".