Fix NVCC/G++ error: casting to int* from void*
The issues arose when trying to compile ArrayFormat.sac in the Stdlib
with NVCC (`-tcuda'). The error message indicated that casting from a
void* to an int* is not allowed - this is not surprising as it's part of
the C++ standard. The problem is why are we needing to do the cast in
the first place, why do we not propagate the type?

The culprit code is this:

  int[.] maxsl1(int[.,.] y)
  {
    z = with {
        ([0] <= iv < take([1], shape(y))) : sel( iv, y);
          } : fold(max, minint());
    return(z);
  }

This can be simplified to:

  noinline
  int[.] maxsl1 ()
  {
    z = with {
        ([0] <= iv < [2]): [1,3,4,6];
          } : fold(max, 0);
    return(z);
  }

(NOTE: needs to be compiled with `-maxwlur 1')

When we look at the code after the optimisation phase (`-bopt') we can
see that we have an int* (AUD) accumulator variable that is then used
for the fold operation. This variable is allocated and assigned a scalar
value at each loop, but is later adjusted (its dimension and shape
increased) such that it matches the dim/shp given in the
WL-assignment field.

Looking through the std.h ALLOC__DATA icms, it became clear that we
intentionally don't allocate the accumulator variable with the correct
type, leading to this problem in the NVCC/G++ world.

Updating the relevant ICMs to use the ALLOC__DATA_BASETYPE ICMs fixes
this issue.