Change CUDA Reg ICMs
Previously we would call cudaHostRegister() for each invocation of
malloc() - this turned out to be very expensive timewise. This commit
changes this by making *only* memory that is about to be sent or recived
from the CUDA device page-locked. In otherwords for each up/down
transfer we use cudaHostRegister()/cudaHostUnregister().