Fix missing module header propagation
This commit extends the serialization/de-serialization traversals
to include the MODULE_HEADERS attribute from N_module. Previously this
was not being done and any function (externally declared) used within
another module or program would result in a CC warning about missing a
prototype, leading to undefined behaviour.

The fix is to generate a function as part of serialization step which
contains the HEADERS (stringset) and that is called during the
de-serialization step.