[HWLOC] Fix syscall redefine when using HWLOC
With mt_barriers we provide our own prototype for syscall(), why is not
entirely clear to me, but in the case of activating HWLOC, we
unintentionally include a prototype of syscall from there. In most
compilers this is not a problem, but when using NVCC we get complaints
that we have redefined syscall().

Additionally this patch removes the execution bit from mt_pth.h