Add DCR subphase after call to eliminate SAA
Sometimes SAA variables are passed as parameters of a function
even though they are not being used. For some reason, dead code
removal is only called about 10 phases later - whether having the
extra unused variables causes a problem or not is not clear, but
it seems safe to removes these earlier.

This patch introduces an additional call to DCR right after the
ESAA phase is done.