Extend Pre/Post function traversal injection and usage
This commit extends the existing pre/post function injection for
traversals by implementing a stack, allowing for a chain of functions to
be called before or after a traversal. Additionally mechanisms have been
implemented that an modify the stack.

We no longer need to - or should - specify a pre/post fun as part
of a traversal's `do' function. Instead, for clarities sake, we
now can add/remove pre/post functions as part of the phase definition
in `phase_*.mac`. This makes it more clear where pre/post funs are
being added and makes it easier to control when one is added or
removed.

It should be noted that there is no automatic removal of pre/post
functions. Users are expected to handle this directly within
`phase_*.mac` - otherwise we'll suffer a leak.

Some concerns:
- pre/post funs are intended to be run at every call of TRAVdo, meaning
  that for each traversal we might have hundreds of calls - given that
  we must iterate through a list *and* check that we can run the fun,
  this _might_ incur a noticeable performance penalty.
- Adding/removing pre/post funs as part of the traversal schema is
  useful, but odd. What I mean is that we do not actually associate
  a pre/post fun to a phase/subphase, but rather a specific traversal.
  This is specified as part of the macro call (we use TR_* names here).
  This could cause confusion for a user expecting to be able to specify
  a phase...