Initial configuration for *new* Master run
Here we add the `.gitlab-ci.yml`, which activates continuous integration
in GitLab for the sac2c project. Through a series of external build
bots (GitLab Runners), we are now able to run build operations
(defined within the `.gitlab-ci.yml`) and generate packages of the
completed builds (these are then stored within GitLab for a finite
amount of time).

The CI operations (the master run) only start up under certain
circumstances, these are:
 * merge-requests: when a user wishes to merge in changes to
   @develop - this **does not** create packages
 * web: user activation within the GitLab WebUI - this creates packages

As of the date of this commit (Nov 2017), GitLab does not have an
internal notion of *merge-request* triggers, meaning that we **cannot**
configure the master run to trigger for merge-requests directly.
The current solution is to use a external service [1] which listens for
events through the GitLab external API - when a merge-request comes it,
it triggers the master run to start.

*NOTE: This commit **only** provides building/packaging of `sac2c`; it
does not do any testing at all.*

*NOTENOTE: Assumes that the build bots are configured with all the
necessary packages.*

[1]: https://github.com/hv15/gitlab-ci-build-on-merge-request