Change test scripts to use CMake defined CC
Previously we were calling `cpp` directly, which is typically fine,
but on some systems (Ubuntu) is only available as a seperate package.
As such, we instead call CMake's found CC and have it do the
preprocessing for us.

NOTE: because we are using CC, we have to explicitly tell it what
      C-language (`-x`) we are using, otherwise it'll not preprocess
      the SaC file as it can't recognise the extension/language.