move loop memory propogation up to deal with CUDA
The CUDA compiler phase introduces the transfer primitives and
alters variables into host and device types. Doing the loop
memory propogation operation during the memory phase doesn't
work with these changes, so the propogation has been moved up.

Its not yet clear if this is 'ok' but some tests show no problems
with this. The key is that we need to have ERCs in place in order
to do it.

Alternatively, if it is the case that doing the propogation this
early causes problems, we investigate what changes need to be made
to the CUDA phases to propogated/update avis' within ther ERCs.