fixed issue with CUDA MLTRAN
MLTRAN is an optimisation that identifies cases where in
loopfun we have a _alloc_ and _host2device_ prf calls before
any WL, and lifts these operations out of the loopfun.

Introducing ERCs caused a slight problem with this as all of a sudden
we have N_id's which refer to vars which do not exist within the current
fundef scope - the checks within MLTRAN (and perhaps else where)
give when when they encounter such a case (as it isn't clear when this
var will be transfered or not).

The primary problem was that within the CUDA WL_memtran phase, we were
not updating the avis of the ERC N_id to point to a host/device specific
avis. This has now been corrected.

Given though that we were dealing with invalid N_id's, the compiler
did not crash because the Mem:Rci (reuse inference) phase identifies
this problem and flushes out the 'invalid' N_id.