Add memory profiling (initial)
We can now profile memory, which currently means collecting the
number of calls to alloc/free and the total allocated size(s).

Unfortunately the profiling feature is in disarray, with portions
of it centric to DISTMEM, which isn't even compiled in. This commit
brings the profiling feature back to working order, but it is unclear
if it is working as specified.

In particular, piece of code from the git history had to be restored
to bring back some of its functionality.