we now always lift allocation (ELMR)
When lifting allocations out of loops, we previously would either:
 a. do a copy operation on an existing parameter to the loopfun
    that matches the correct type/shape.
 b. or create a new variable with the same type/shape.

This works fine, from a schematic point of view. Practically though,
it means allocating *and* copying the content of one buffer into
another, when by-design that buffer is going to be write to (not
read).

For various benchmarks, this led to in some case an extreme increase
in the runtime of the application. Primarily because of the the copy
operations, but also because of the various RT check doing
alloc_or_reuse for the buffer.

Other than convience, there is not really reason to *not* create a
new variable anyway - as such this commit removes point (a). Any time
we wish to lift an allocation out of the loop body, we create a new
variable with the correct type/shape and allocate this (in ELMR)