[version manager] move to argparse
We have specified through CMake that python 2.7.* is now required.
This move is so that we can use argparse - which replaces optparse
which is depreciated in python since 2.7. The primary reason to move
is that argparse includes further checking of arguments and flags
which we currently manually check for. This has helped to simplify
some of the code a bit.

Additionally we have replaced the `command' parsing with argparse
directly, again helping to simplify the code.

Further changes include minor spelling fixes. No other changes to
the code are done.