[version manager] support arbitrary creation of sac2c symlinks
We now support creating symlinks arbitrarily - this means that we
don't need to inspect the filesystem for installed version of sac2c.
This is very useful for the build system, whereby we can create symlinks
with this tool instead of cmake, giving us better control.

Potentially this can be used for creating symlinks to a local build
directory.