redo/extend symlinking during `make install`
Previously, we used `cmake -E` internal symlink mechanism to create
links. We now use the sac2c-version-manager script instead, as this
is slightly more reliable when it comes to 'installing' different
links.

This is also multi-platform compatible, though for Windows and
forwards slashes this is untested - Python os.* supports Windows so
it should work, or otherwise gracefully fail.