[version manager] minor fixes to documentation
In particular the install command documentation has been changes to be
more clear. Other commands have also been changed slightly.

Additionally we've cleaned up the code a bit and have included some more
CMake variables from the configuration phase in the script - we used to
infer these but it is likely better to rely on the build system for this
information.