[phm] turn it on again per default for seq
It is unclear why this was ever off - several tests show that the
PHM works fine for the seq target. Via sac2crc it was only activated
for the mt* targets.

This commit changes the sac2crc files to activate PHM for seq and mt*
targets - cuda* targets have been explicitly excluded from this.

This is chance this might bork on MacOS...