GCC 8 warning fixes
Thie commit resolves some of the new warnings that GCC 8 issues.

Specifically warnings `-Wcast-function-type` and `-Wrestrict`. The first
clearly deals with casting from one function type to another and the
latter deals with breaks in `restrict`, that is we have _unsafe_
operations, e.g. reading and writing using the same pointer.