[emr] major fixes (prf modarray and arg order)
This commit provides two major fixes to the EMR related traversals.

The first deals with the _idx_modarray_* primitives which from
observance always* reuse the array that they reference as arguments.
The EMR traversal did *not* take this into account, meaning that we
forced memory to be re-allocated anew *and* copy the data. This clearly
can lead to performance problems. The currently solution is to filter
out all of the referenced arrays from the ERCs - we can loose a large
number of these.

The second part deals with a small mistake in the order in which new
arguments are added to the recursive-loop call - the function that finds
suitable candidates returned these in the wrong order, leading to code
generation problems. This is critical if we lift more than one
allocation out of a loop body.

Also a minor improvement for identifying good ERC candidates - we check
more then just the type for shape information.