[profile] various changes for clarity
The profiling system is clearly built around timing, but with the
addition of the memory profiling, it has become clear that the names
of some macros and types is unclear (and even confusing).

At its core, the profiling system keeps a simple linked-list of
'records' which are used to store some function and with-level
attributes, such as the internal ID of the function and the number of
times it is called. This record holds *no* timing information, instead
this information is stored within a global array which is indexed to
match the given function ID and appplication count. This record however
is called the `timer_record`, a name which clearly does not reflect what
it is actually doing. This is been updated.

With this commit we now collect memory profiling information at
a per-function granularity. This makes it easier to see where memory
operations are actually happening.

Additionaly some shuffling of macros was done. This shuffle
re-introduces the record struct whenever we do profiling.
This is important as the record struct stores the previous function ID
and application number, which we need in function and memory profiling.