Further fixes to ERC filtering
This commit resolves (1) a minor problem with FRC that led to improper
filtering of RCs (we forgot to traverse through N_fundef ERCs, which
caused all sorts of problems). (2) we now do a further stage of
filtering that removes all RCs from being an ERC in a later WL
application. There are instances of programs where for example we can
have:

    a1 = with {
          ...
        } : modarray(a); /* RC: a */

    b1 = with {
          ...
        } : modarray(b); /* ERC: a */

When this occurs, the filtering previously would remove `a` from both
the RC and ERC fields in both WL, meaning we kill off all opportunities
for memory reuse.

A new filtering step was added that specifically removes this case,
thereby resolving the killing of RCs - the result then would now be:

    a1 = with {
          ...
        } : modarray(a); /* RC: a */

    b1 = with {
          ...
        } : modarray(b);

If done in a loop, we now correctly identify an opertunnity to lift out
the allocation for the second WL. In such a case, after the MEM stage,
we end up with:

    a_mem = a;
    a1 = with {
          ...
        } : modarray(a, MEM: a_mem);

    b_mem = b_lift_tmp;
    b1 = with {
          ...
        } : modarray(b, MEM: b_mem);

    ...

    A, ... = loop_0 (a1, b1, a_mem, b_mem);