[vers manager] minor fix to support python2.6
On CentOS 6.*, Python2.6 is the primary variant of python available
through the package manager. Unfortunately, it is very cumbersome to
install a newer version of Python, as several systems tools (such as
yum) have a hard dependency on the older version. Because of this we
run into compatibility problems:

Python2.6 does not support dict comprehensions (these were introduced in
python2.7 - see [PEP274][1]). We use this in two places within the
version script. This patch changes the comprehensions into normal loop
operations. This works fine with later versions of Python2.* branch.

It's unfortunate that we need to worry about this given that python2.6
is no longer support by the Python devs... perhaps we should rethink
out position on supporting CentOS 6.

[1]: https://www.python.org/dev/peps/pep-0274/