[emr] use existing RC/ERCs for alloc loop lift
I have found several cases where, when doing an allocation lift out of a
loop, we have no variable to add to the args of the recursive call. The
primary reason for why this happens is that the WL alloc which we are
lifting might have its LHS assignment be used as an RC/ERC in a later
WL. There are several things we could do, like when we know we are going
to lift an allocation, we remove the LHS from all RC/ERC references.
Doing would intentionally remove a case for reuse - which is could
defeat the purpose of lifting out an allocation at all.

Instead this commit implements a different solution - given the
previously described case, it is clear that we might end up with
multiple candidates for each WL's RC/ERC. Its highly unlikely that we
would use all (even one of them, but so long as there is one, we can be
fairly sure that at worst, it'll be reused at runtime). Given this, we
can just *leech* a candidate from one of the WL RC/ERCs for use as an
arg at the recursive call. We just need to make sure that we take from a
RC/ERC list that has more than one candidate. Later traversals like RI
will filter these out when it finds that the N_avis is being referenced.

To be clear, this new facility is *only* used as a last resort, if we
have candidates within the N_fundef ERC list, this will never get used.