dynamically update build targets at configure time
This commit introduces the notion that at configure time we can add to
(and potentially remove from) the build targets list for the runtime
libraries depending on if certain feature are on or off.

We do this now in the case of CUDA, as this feature needs now RT libs
compiled for it.