[TCcnt] Fixed warnings from size_t TCcountExprs and refactored 2 funcs
Where SAC input code(PMvar, COconst2Int,etc.) is needed for size_t args,
used DBUG assertions and left warnings in place to trace back later
when new data type gets introduced.

Refactored TClookupIdsNode again and LFUindexOfMemberIds to use bool
as both are searching for whether a node is a member of a chain, and
it is used as a true or false for the node being a member.
Added/Changed comment to explain the change.