Use system C-compiler to build runtime libraries
When building the CUDA SAC runtime libraries, we were encountering
problems with warnings (and sometimes errors) being issues by NVCC. This
is understandable as some parts of the runtime library code is *not* C++
compliant.

It isn't nessicary to compile the CUDA SAC runtime libraries with NVCC
as there are no kernel launches. As such we can safely use the system C
compiler to do this.

This commit extends sac2crc syntax slightly by encoding a
runtime-library specific COMPILE command which is used now when building
the SAC compiler.