Fix runtime libsac test dependencies
Previously whenever we wanted to just compile a runtime test, we would
often have dependency resolving problems because the runtime libraries
are built as an ExternalProject.

We now expose the libsac-<target> targets and use these for dependency
handling.

These targets are not LIBRARY targets, and so cannot be directly used
with TARGET_LINK_LIBRARIES. Additionally, the target is a utility
target, which has no properties. We add some properties to expose both
the library build path and the rpath, which we can explicitly use to
correctly link the library.