fix cmake/make parallel build warning
In some instances when building in parallel (e.g. `-jN`), targets that
use a `${CMAKE_COMMAND}` call will causes make warnings:

   warning: jobserver unavailable: using -j1. Add `+' to parent make rule.

This indicates that we have not passed on the jobserver environment to
the sub-makefile Make calls.

The solution in this instance is rather trivial, we add `+` to the
`${CMAKE_COMMAND}` calls which furthermore allows us to build in
parallel in the sub-makefile Make calls.