Use unistd.h header for syscall()
Previously we declared this ourselves, leading to compiler warnings
(especially with NVCC).

Additionally we remove the `-DPIC` from compiler CFLAGS as this is
redundant. At some point there was a need to specify `#ifdef PIC` in
source files because of how libtool behaved (see GNU mailing lists for
details), this is no longer the case.

Finally we make a minor change to the how we propogate the definitions
to the C-compiler, specifically we now use the SET_COMPILE_OPTIONS
macro, as this is dictated by future release of CMake. We also treat the
flags as lists instead of strings, as ADD_DEFINITIONS can't handle a
string of flags very well, it quotes the whole thing leaving the
compiler to cough...