add cpp macro overloading-like support
We define a CPP macro `MACRO_ARGCOUNT` which counts the number of
arguments passed. This is useful because it allows us to define macros
for different argument counts, by hiding the specific macro behind a
generic parent macro, e.g.

```
TEST(...) ---> TEST0 () "no args passed"
          |
          ---> TEST1 (a) a
          |
          ---> TEST2 (a, b) b
          |
         etc.
```

We define TEST such that for zero arguments we return a string, for
one argument we return it, etc. In this sense, we achieve a form of
overloading at the level of cpp macros. Further details of usage are in
`argcount-cpp.h`.

This is currently being used within `types.h` for the `PHASENAME` macro,
which constructs a string made of the arguments in different orders
depending on the number of arguments passed.

NOTE, we only support up to 64 arguments maximum, but this can be
extended to support more.