fix incorrect CPACK version number
Over the summer of 2018, we switched to a full semantic version number but
forgot to update the macro responsible for deriving the version number
components. This means that all packages in this time period have the
wrong internal version, meaning all dependendency checks fail.

Additionally we extend the masterrun to also _install_ build packages as
an additional sanity check.