Add GTest as an external project
Instead of relying on the system to have GTest installed, we download
and compile this ourselves via the ExternalProject macro.

We do this as there is some evidence that GTest is not consistently
packaged. This means that in some cases it's not clear if CMake's
FindGTest will work and correctly propogate flags when linking. For
instance on CentOS 6 GTest is compiled with pthread support, but
FindGTest does not pick this up, leading to undefined symbols error
when linking. Additionally, there has been some back and forth on
Ubuntu and Debian regarding how to package that thing [1].

Loading/compiling/linking like this is supported by the GTest
developers, see [2].

We explicitly avoid compiling/linking againt GMock as it has strong
dependencies on C++11 support. We currently download GTesT 1.8.1 which
is the last version to support non-C++11 supported compilers.

Additionally we move to 3.10 as the minimum version to get access to
some more fancy GTest related macros (via CMake GoogleTest include).

[1]: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=802587
[2]: https://github.com/google/googletest/blob/master/googletest/README.md