fix some build flags (macro vardic warning)
When compiling the production version of sac2c, we used `std=c99`
instead of `std=gnu99`, which led to interesting warnings regarding
vardic macros.

Similarly we found the same problem with CLANG (on Apple), and fix this.

Additionally we have stopped using `-Wpendatic` for versions of GCC <=
5. This is because there are further warnings issued that can't be
disabled (about macro arguments not being used when using varargs).
Unfortunatly disabling the warning using `#pragma GCC diagnostic` also
doesn't work on macros - theres a bug report for this:
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=53431