add EMRTU optimistation
This optimisation works together with the CUDA IWLMEM optimisation to introduce
CUDA memory transfers into the AST. We need this traversal when using the EMR
optimisation as IWLMEM would otherwise lead to inoptimal code generation.

IWLMEM, when encountering a lift ERC, would within the loop function create a
host2device and cause a memory allocation and free on the CUDA device. The
EMRTU (EMR Type Update) traversal identifies such cases, and correclty
transforms the lifted ERCs into CUDA device_types *and* fixes the arguments of
the loop function applications (initial and recursi