fix spurious CMake build problem
This is a known issue with running a CMake built Makefile system
that eventually calls and External_Project. The various _steps_ in
the External_Project are **not** considered individual jobs with
interdependencies. This can therefor cause multiple instances of one
_step_ to be run, leading to undefined behavior. The issue is
documented in [1] and affects all versions of CMake > 3.10.

The workaround for this is to explicitly set dependencies on the _steps_
and remove DEPENDS from the External_Project.

[1]: https://gitlab.kitware.com/cmake/cmake/issues/18663