add dummy cuda call to initilise context
depending on the device/driver version/and system configuration, it
can take variable amount of time for the device to be initialised and
the CUDA context to be created *before* any communication or kernels
are launched. Ordinarily this happens on the first API call (in most
SaC cases a cudaMalloc or kernal invocation) which can hide the
actual overhead.

This commit adds a dummy function call as part of the CUDA startup
ICM which should _capture_ the overhead cost, such that runtime
experiments can excludes these.