properly perform CUDA page-locking (pinning)
This commit introduces a large number of changes which adds proper
(this means better) support for page-locking memory for CUDA
asynchronous mmory transfers.

Previously, we had naively page-locked memory immeditely before the
transfer is performed. In cases where the transfer occurs in a loop, and
it cannot be lifted out, this can negatively affect performance.

By way of a traversal, we mark those N_avis which will eventually
(aliasing) be an argument to a CUDA H2D or D2H operation. With this, we
extend the allocation ICMs to perform the page-locking at the point of
initial allocation.

To give a more concrete idea of how this looks, previously we did:

    ...
    a = alloc ()
    ... do something with a ...
    a_dev = alloc ()
    pin (a)
    cuda_memcpy (a, a_dev)
    unpin (a)
    free (a)
    ...

to

    ...
    a = alloc () [PIN]
    pin (a)
    ... do something with a ...
    a_dev = alloc () [NOPIN] // device memory
    cuda_memcpy (a, a_dev)
    unpin (a)
    free (a)

This commit is not READY, though it seems to work all right for my
benchmarks - I have a sneaking suspecion that this will not work for
modules. Other remarks are below.

TODO:
 - unclear if this works with compiling modules
 - SACarg setup concerns me (as well as codegen-ND_DECL__MIRROR_EXTERN)
   as its not clear to me if we want to encapsulate pinning information
   or not.
 - Before pinning or unpinning we check the array's descriptor - this
   means that for each pinned array, we add two conditionals to the
   program... pinning state is statically determined though, so maybe
   C-compile optimises the conditional away?