nasty hack
previous commit did not consider nested N_fundef, meaning that we
missed out on propograting the pinned flag from LHS to RHS.

This commit provides a hackish way of doing this, in that we scan up and
then go back down the assign chain.

I'm leaving it there for code changes as I'm running out of time. The
descriptor and runtime headers is enough to manually tag the appropriate
variables that should be pinned, but where missed.

TODO:
 - encountered interesting problem with with2 generated stuff