disable pedantic warnings for the argcount header
We make use of a GCC system header marker to indicate that the
headerfile is not C standard compliant, and thus disable the warnings
relating to zero-argument variadic macros.

This commit also fixes a small mistake in config.cmake where we added
the no warning flag to the wrong var (oops).