add lazy pinning
Rather then trying to infer which avis needs to be pinned, which also
means having to handling aliasing, we instead use the CUDA transfer as
an indication for memory that needs to be pinned, and pin the memory
then and there. Furthermore we mark the memory (in the descriptor) as
being pinned, meaning further calls to pin are skipped over. We then
unpin when we free.

This is achieved through the ICMs. Unfortunately we introduce several
more conditionals per CUDA transfer and on every free.