initial commit of CUDA async delay
We discovered a problem with intermittent race-conditions caused by the
introduction of CUDA async memcpys into the code. Funnily, due to the
design of used benchmarks and examples codes, the race-conditions where
hidden by the inherent latency of the underlying operations. However a
race condition can be forced, using for instance the below program:

```sac
use Array: all;

noinline
int foo( int x) {
  return x;
}

int main() {
  a = with {
        (. <= [i] <= .) : foo( i);
      } : genarray( [1000], 0);

	ArrayIO::print( a);

	do {
    a = a+ 1;
  } while( a[0] < 20);

	ArrayIO::print( a);
  return 0;
}
```

Compiling with `-doEMRCI -t cuda -bmem`, we see that prints of `a`
are correct. If we compile using target `cuda_reg` (or `cuda_alloc`)
the printing of `a` shows that we have successfully written to `a` before
it was done transferring.

This race-condition exists because in CUDA, most CUDA operations are
non-blocking on the host *and* are scheduled separately. The above code
demonstrates this. In order to fix this we need to introduce some kind
of barrier/sync. In the context of code-generation, and maintaining the
data-flow of our IR, this is non-trivial.

So far, our solution to this is to introduce some latency, whereby we
delay memory accesses on the host as late as possible. Additionally,
we introduce a synchronisation point (which as far as we can determine
is completely necessary). That being said, if we succeed in increasing
the latency to a point, the synchronise could become a noop and not
incur a performance penalty.

We do the delay by extending the F_host2device/F_device2host primitives
to have _start_ and _end_ points. The start point is when the transfer
is meant to start (from the perspective of the host), and the end point
is where we need to synchronise before we access memory. In the case of
host2device, we want to start as soon as possible, so we will need to
push the _start_ primitive up the dataflow graph. Conversely, for
device2host, we need the _end_ point to be pushed down.

This commit introduce the translation from the single primitive to it's
pair primitive. Later commits in the serious will introduce mechanisms
to push the start/end points.