add various sync modes for CUDA async methods
With this commit we add four modes for synchronising on memory transfers
when using CUDA async transfer mechanisms. These are NONE: we do not
synchronise (current model), DEVICE which uses the device synchronise
primitive, STREAM wherein we sychronise on the stream, and finally
CALLBACK which is similar to STREAM, but does uses spinlocks instead of
CUDA event system.