fix argument traversal in N_with
Previously, we did not detect if a given N_avis was on the RHS within
a N_with code-block. This lead to compiler errors in WLT phase. We fix
this by not capturing N_assign from within a N_with into our LUT.
Instead, in such a case, we add the next_assign of the N_with.