[emr-lift] change finding of N_avis for rec-loop
When lifting out a memory allocation, we need to find a suitable
candidate for the recurisve loop function application to match the same
shape/type of the new argument. We take this from the N_fundef ERC list,
previously filled in by the EMRCI traversal. We scan through the
candidates from first in the chain to last, which matchs with their
position in the code bottom to top. The rational for this was to
maximise the reuse capability of these candidates.

We change this matching around, and now search from last for first in
the chain. The rational for this is, (a) we already filter through the
N_fundef ERC list such that it contains no referenced ERC (associated
with a WL) and (b) in the case of CUDA, we are better to handle the case
where *all* h2d transfers are lifted out of a inner loop.