[hwloc] add support for hwloc v2
Actually there isn't much we need to change, as much of our HWLOC code
deals with binding to PUs/CPUs. The main changes in hwloc v2 mainly
affect how to bind based on cache/memory locality, which we don't use
here. Further changes involve a change in how topology is orgonised, for
instance NUMA nodes are no longer treated as logical/real things,
instead they are treated as _memory children_ instead. The immediate
affect is that taversing the topology is different. Further
orgonisational changes to the topology exist, but really are handled by
the API such that we need only perform minor changes to the existing
code.