[emr] fix turning on EMR by default
In the filter traversal (emrcf, which aliases to frc) we traverse into
the N_break node (with-op associated with N_fold). Normally there is
some MEM associated with the node, but this only happens once we reach
mem phase (mem:alloc). As we perform filtering earlier in the opt phase,
we run into NULL nodes.

This commit fixes this by adding a check...